rename pascal case and kebab case to follow camel case
