+++ src/sections/safeStaking.tsx
